WIP: put_and_get
[x] completed failing tests
[x] renamed FileSystemCache to FileSystemStorage
[x] extract functions
[ ] ensure OO class has 1:1 implementation
[ ] extract class hierarchy
[ ] ensure test are implemented correctly for new target structure

-– [x] describe
-– [x] fail
-> [ ] pass
