+++ storage/file_system/__init__.py
@@ -1,5 +1,2 @@
-from io import BufferedReader, BufferedWriter
-from typing import TypeVar, Union
 
-Text = TypeVar('Text', Union[str, bytes])
-BufferedIO = Union[BufferedWriter, BufferedReader]
+
